pyproject.toml
[tool.poetry]
name = "blog_proto"
version = "0.1.0"
description = "Blog framework/prototype"
authors = ["Mustafa Prodhan <mustafa.prodhan.317@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.1.7"
Pillow = "^8.1.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
